add state generator
Needs more test cases and also a clean way to check if a state is in the
list of generated states. We don't want to rely on the order of the
states in the list, because the order doesn't matter for the actual
program. Therefore it would be unwise to expect a certain order in the
testcode.