fix an infitive regress for input "1 2 3 4 5 6 7 8 0"
This fixes an edge case. If the given input is already the final state
of the game, the shortest-path function would never terminate. It would
try to get a value from the empty solution space, that value would be
nil, and then it would add nil to the solution path and repeat forever.